requirements.txt
# Python Package Requirements
# Compatible with Python 3.11, 3.12, and 3.13
#
# IMPORTANT: Using the NEW google-genai package (the old google-generativeai is deprecated)

google-genai>=0.2.0
streamlit>=1.28.0
oracledb>=2.0.0
python-dotenv>=1.0.0
